vinayaka_requirements_docker.txt
Package                  Version     
------------------------ ------------
absl-py                  1.4.0       
addict                   2.4.0       
aiofiles                 22.1.0      
aiosqlite                0.19.0      
ansi2html                1.8.0       
antlr4-python3-runtime   4.8         
anyio                    3.6.2       
appdirs                  1.4.4       
argon2-cffi              21.3.0      
argon2-cffi-bindings     21.2.0      
asttokens                2.2.1       
attrs                    23.1.0      
Babel                    2.12.1      
backcall                 0.2.0       
beautifulsoup4           4.12.2      
bleach                   6.0.0       
cachetools               5.3.0       
certifi                  2022.12.7   
cffi                     1.15.1      
charset-normalizer       3.1.0       
click                    8.1.3       
comm                     0.1.3       
contourpy                1.0.7       
cycler                   0.11.0      
dash                     2.9.3       
dash-core-components     2.0.0       
dash-html-components     2.0.0       
dash-table               5.0.0       
dbus-python              1.2.16      
debugpy                  1.6.7       
decorator                5.1.1       
defusedxml               0.7.1       
deprecation              2.1.0       
docker-pycreds           0.4.0       
executing                1.2.0       
faiss-gpu                1.6.5       
fastjsonschema           2.16.3      
filelock                 3.12.0      
Flask                    2.2.3       
fonttools                4.39.3      
gdown                    4.7.1       
gitdb                    4.0.10      
GitPython                3.1.31      
google-auth              2.17.3      
google-auth-oauthlib     1.0.0       
googledrivedownloader    0.4         
grpcio                   1.54.0      
h5py                     3.8.0       
hydra-core               1.1.0       
idna                     3.4         
imageio                  2.28.0      
importlib-metadata       6.6.0       
importlib-resources      5.12.0      
ipykernel                6.22.0      
ipython                  8.12.0      
ipython-genutils         0.2.0       
ipywidgets               8.0.6       
isodate                  0.6.1       
itsdangerous             2.1.2       
jedi                     0.18.2      
Jinja2                   3.1.2       
joblib                   1.2.0       
json5                    0.9.11      
jsonpatch                1.32        
jsonpointer              2.3         
jsonschema               4.17.3      
jupyter-client           8.2.0       
jupyter-core             5.3.0       
jupyter-dash             0.4.2       
jupyter-events           0.6.3       
jupyter-packaging        0.12.3      
jupyter-server           2.5.0       
jupyter-server-fileid    0.9.0       
jupyter-server-terminals 0.4.4       
jupyter-server-ydoc      0.8.0       
jupyter-ydoc             0.2.4       
jupyterlab               3.6.3       
jupyterlab-pygments      0.2.2       
jupyterlab-server        2.22.1      
jupyterlab-widgets       3.0.7       
kiwisolver               1.4.4       
llvmlite                 0.39.1      
Markdown                 3.4.3       
MarkupSafe               2.1.2       
matplotlib               3.7.1       
matplotlib-inline        0.1.6       
MinkowskiEngine          0.5.4       
mistune                  2.0.5       
mit-semseg               1.0.0       
nbclassic                0.5.5       
nbclient                 0.7.3       
nbconvert                7.3.1       
nbformat                 5.8.0       
nest-asyncio             1.5.6       
networkx                 3.1         
notebook                 6.5.4       
notebook-shim            0.2.2       
numba                    0.56.4      
numpy                    1.23.5      
oauthlib                 3.2.2       
omegaconf                2.1.2       
open3d                   0.13.0      
opencv-python            4.7.0.72    
packaging                23.1        
pandas                   2.0.1       
pandocfilters            1.5.0       
parso                    0.8.3       
pathtools                0.1.2       
pexpect                  4.8.0       
pickleshare              0.7.5       
Pillow                   9.5.0       
pip                      20.0.2      
pkgutil-resolve-name     1.3.10      
platformdirs             3.2.0       
plotly                   5.4.0       
plyfile                  0.9         
prometheus-client        0.16.0      
prompt-toolkit           3.0.38      
protobuf                 4.22.3      
psutil                   5.9.5       
ptyprocess               0.7.0       
pure-eval                0.2.2       
pyasn1                   0.5.0       
pyasn1-modules           0.3.0       
pycparser                2.21        
Pygments                 2.15.1      
PyGObject                3.36.0      
pykeops                  1.4.2       
pyparsing                3.0.9       
pypng                    0.20220715.0
pyrsistent               0.19.3      
PySocks                  1.7.1       
python-dateutil          2.8.2       
python-json-logger       2.0.7       
python-louvain           0.16        
pytorch-metric-learning  2.1.0       
pytz                     2023.3      
PyYAML                   6.0         
pyzmq                    25.0.2      
rdflib                   6.3.2       
requests                 2.28.2      
requests-oauthlib        1.3.1       
retrying                 1.3.4       
rfc3339-validator        0.1.4       
rfc3986-validator        0.1.1       
rsa                      4.9         
scikit-learn             1.2.2       
scipy                    1.10.1      
seaborn                  0.12.2      
Send2Trash               1.8.0       
sentry-sdk               1.20.0      
setproctitle             1.3.2       
setuptools               45.2.0      
six                      1.16.0      
smmap                    5.0.0       
sniffio                  1.3.0       
soupsieve                2.4.1       
stack-data               0.6.2       
tenacity                 8.2.2       
tensorboard              2.12.2      
tensorboard-data-server  0.7.0       
tensorboard-plugin-wit   1.8.1       
terminado                0.17.1      
threadpoolctl            3.1.0       
tinycss2                 1.2.1       
tomli                    2.0.1       
tomlkit                  0.11.7      
torch                    1.7.1+cu110 
torch-cluster            1.5.9       
torch-geometric          1.7.2       
torch-points-kernels     0.6.10      
torch-scatter            2.0.7       
torch-sparse             0.6.9       
torch-spline-conv        1.2.1       
torchnet                 0.0.4       
torchsparse              1.1.0       
torchvision              0.8.2+cu110 
tornado                  6.3.1       
tqdm                     4.65.0      
traitlets                5.9.0       
typing-extensions        4.5.0       
tzdata                   2023.3      
urllib3                  1.26.15     
visdom                   0.2.4       
wandb                    0.15.0      
wcwidth                  0.2.6       
webencodings             0.5.1       
websocket-client         1.5.1       
Werkzeug                 2.2.3       
wheel                    0.40.0      
widgetsnbextension       4.0.7       
y-py                     0.5.9       
yacs                     0.1.8       
ypy-websocket            0.8.4       
zipp                     3.15.0      
